# nodeGIS/api.yaml
openapi: 3.1.0
info:
  title: 默认模块
  description: ''
  version: 1.0.0
...
paths:
  /auth/register:
    post:
      summary: 用户注册
      deprecated: false
      description: ''
      tags:
        - 认证
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example: ''
      responses:
        '200':
          x-apifox-name: 注册成功
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    examples:
                      - uuid-user-123
                  message:
                    type: string
                    examples:
                      - 注册成功
                x-apifox-orders:
                  - userId
                  - message
                x-apifox-ignore-properties: []
          headers: {}
          x-apifox-ordering: 0
        '400':
          x-apifox-name: 无效请求（例如：参数验证失败）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: 输入参数无效
                details:
                  - field: email
                    issue: 邮箱格式不正确
          headers: {}
          x-apifox-ordering: 1
        '409':
          x-apifox-name: 资源冲突（例如：邮箱或用户名已被注册）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: USER_ALREADY_EXISTS
                message: 该邮箱已被注册
          headers: {}
          x-apifox-ordering: 2
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 3
      security: []
      x-apifox-folder: 认证
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-372255083-run
  /auth/login:
    post:
      summary: 用户登录
      deprecated: false
      description: ''
      tags:
        - 认证
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              identity: '100'
              password: Excepteur enim sed
      responses:
        '200':
          x-apifox-name: 登录成功
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers: {}
          x-apifox-ordering: 0
        '400':
          x-apifox-name: 无效请求（例如：缺少 identity 或 password）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: 缺少 password 字段
          headers: {}
          x-apifox-ordering: 1
        '401':
          x-apifox-name: 认证失败（例如：用户名或密码错误）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INVALID_CREDENTIALS
                message: 用户名或密码错误
          headers: {}
          x-apifox-ordering: 2
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 3
      security: []
      x-apifox-folder: 认证
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-372255084-run
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - phone
        - password
      properties:
        username:
          type: string
          description: 用户名
        email:
          type: string
          format: email
          description: 邮箱
        phone:
          type: string
          description: 手机号
        password:
          type: string
          format: password
      x-apifox-orders:
        - username
        - email
        - phone
        - password
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    LoginRequest:
      type: object
      properties:
        identity:
          type: string
          description: 手机号/用户名/邮箱
        password:
          type: string
          format: password
      x-apifox-orders:
        - identity
        - password
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Token
        expiresIn:
          type: integer
          description: 过期时间
        user:
          type: object
          properties:
            userId:
              type: string
              description: 用户ID
            username:
              type: string
              description: 用户名
          x-apifox-orders:
            - userId
            - username
          x-apifox-ignore-properties: []
      x-apifox-orders:
        - token
        - expiresIn
        - user
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 业务错误码，便于前端处理
          examples:
            - VALIDATION_ERROR
        message:
          type: string
          description: 人类可读的错误信息
          examples:
            - 输入参数无效
        details:
          type: array
          description: （可选）详细的验证错误列表
          items:
            type: object
            properties:
              field:
                type: string
                examples:
                  - email
              issue:
                type: string
                examples:
                  - 邮箱格式不正确
            x-apifox-orders:
              - field
              - issue
            x-apifox-ignore-properties: []
      required:
        - code
        - message
      x-apifox-orders:
        - code
        - message
        - details
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-apifox:
        addTokenTo: header
        headerPrefix: Bearer
servers: []
security:
  - jwt: []
    x-apifox:
      required: true
      schemeGroups:
        - id: 2E3vUtLwdMRJwtG6NLo0V
          schemeIds:
            - jwt
      use:
        id: 2E3vUtLwdMRJwtG6NLo0V
        configs:
          jwt:
            authConfigs:
              x-apifox:
                token: >-
                  eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw